requirements.txt
django
face_recognition
djangorestframework